+++ st2actions/st2actions/notifier/notifier.py
@@ -19,7 +19,9 @@ from kombu import Connection
 from oslo_config import cfg
 
 from st2common import log as logging
-from st2common.constants import action as action_constants
+from st2common.constants.action import LIVEACTION_STATUS_SUCCEEDED
+from st2common.constants.action import LIVEACTION_FAILED_STATES
+from st2common.constants.action import LIVEACTION_COMPLETED_STATES
 from st2common.constants.triggers import INTERNAL_TRIGGER_TYPES
 from st2common.models.api.trace import TraceContext
 from st2common.models.db.liveaction import LiveActionDB
@@ -42,12 +44,6 @@ LOG = logging.getLogger(__name__)
 
 ACTIONUPDATE_WORK_Q = liveaction.get_queue('st2.notifiers.work',
                                            routing_key=publishers.UPDATE_RK)
-ACTION_COMPLETE_STATES = [
-    action_constants.LIVEACTION_STATUS_SUCCEEDED,
-    action_constants.LIVEACTION_STATUS_FAILED,
-    action_constants.LIVEACTION_STATUS_TIMED_OUT,
-    action_constants.LIVEACTION_STATUS_CANCELED
-]
 
 ACTION_SENSOR_ENABLED = cfg.CONF.action_sensor.enable
 # XXX: Fix this nasty positional dependency.
@@ -73,7 +69,7 @@ class Notifier(consumers.MessageHandler):
         extra = {'live_action_db': liveaction}
         LOG.debug('Processing liveaction %s', live_action_id, extra=extra)
 
-        if liveaction.status not in ACTION_COMPLETE_STATES:
+        if liveaction.status not in LIVEACTION_COMPLETED_STATES:
             LOG.debug('Skipping processing of liveaction %s since it\'s not in a completed state' %
                       (live_action_id), extra=extra)
             return
@@ -111,12 +107,12 @@ class Notifier(consumers.MessageHandler):
                 liveaction=liveaction, execution_id=execution_id,
                 notify_subsection=notify.on_complete,
                 default_message_suffix='completed.')
-        if liveaction.status == action_constants.LIVEACTION_STATUS_SUCCEEDED and notify.on_success:
+        if liveaction.status == LIVEACTION_STATUS_SUCCEEDED and notify.on_success:
             self._post_notify_subsection_triggers(
                 liveaction=liveaction, execution_id=execution_id,
                 notify_subsection=notify.on_success,
                 default_message_suffix='succeeded.')
-        if liveaction.status == action_constants.LIVEACTION_STATUS_FAILED and notify.on_failure:
+        if liveaction.status in LIVEACTION_FAILED_STATES and notify.on_failure:
             self._post_notify_subsection_triggers(
                 liveaction=liveaction, execution_id=execution_id,
                 notify_subsection=notify.on_failure,