+++ st2common/st2common/services/action.py
@@ -171,7 +171,7 @@ def request_cancellation(liveaction, requester):
     if liveaction.status == action_constants.LIVEACTION_STATUS_CANCELING:
         return liveaction
 
-    if liveaction.status not in action_constants.CANCELABLE_STATES:
+    if liveaction.status not in action_constants.LIVEACTION_CANCELABLE_STATES:
         raise Exception('Unable to cancel execution because it is already in a completed state.')
 
     result = {